+++ api/v1/server/run/run.go
@@ -43,6 +43,7 @@ import (
 	hatchetmiddleware "github.com/hatchet-dev/hatchet/api/v1/server/middleware"
 	"github.com/hatchet-dev/hatchet/api/v1/server/middleware/populator"
 	"github.com/hatchet-dev/hatchet/api/v1/server/middleware/ratelimit"
+	"github.com/hatchet-dev/hatchet/api/v1/server/middleware/telemetry"
 	"github.com/hatchet-dev/hatchet/api/v1/server/oas/gen"
 	"github.com/hatchet-dev/hatchet/pkg/config/server"
 	"github.com/hatchet-dev/hatchet/pkg/repository/postgres/dbsqlc"
@@ -563,12 +564,14 @@ func (t *APIServer) registerSpec(g *echo.Group, spec *openapi3.T) (*populator.Po
 	})
 
 	rateLimitMW := ratelimit.NewRateLimitMiddleware(t.config, spec)
+	otelMW := telemetry.NewOTelMiddleware(t.config)
 
 	// register echo middleware
 	g.Use(
 		loggerMiddleware,
 		middleware.Recover(),
 		rateLimitMW.Middleware(),
+		otelMW.Middleware(),
 		allHatchetMiddleware,
 	)