+++ sdks/python/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to Hatchet's Python SDK will be documented in this changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21.2] - 2025-11-13
+
+### Added
+
+- Adds an OTel option to allow you to include the action name in the root span name for task runs.
+
+### Changed
+
+- Span kinds (e.g. producer, consumer) have been added to OpenTelemetry spans created by the SDK to better reflect their roles.
+
 ## [1.21.1] - 2025-11-08
 
 ### Changed